[step3]refactor : coverImageEntity 제거

+++ src/main/java/nextstep/courses/domain/cover/CoverImageRepository.java
@@ -1,7 +1,7 @@
 package nextstep.courses.domain.cover;
 
 public interface CoverImageRepository {
-    int save(CoverImage coverImage);
+    Long save(CoverImage coverImage);
 
     CoverImage findById(Long id);
 }
